+++ src/codecs/hash.rs
@@ -98,7 +98,7 @@ impl Crc32 {
 			let a = (self.crc ^ byte as u32) & 0xFF;
 			let b = self.crc >> 8;
 
-			self.crc = CRC_TABLE[a as uint] as u32 ^ b;
+			self.crc = CRC_TABLE[a as uint] ^ b;
 		}
 	}
